requirements.txt
Flask==3.1.1
flask_login
python-dotenv==1.0.0
pytest
boto3
pandas
numpy
bs4
gunicorn
gevent
google-api-python-client
google-auth
matplotlib
